+++ conf/cors.js
@@ -1,4 +1,5 @@
-const app = require('express');
+const express = require('express');
+const app = express();
 const cors = require('cors');
 
 module.exports = app.use(cors({
